+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Changed `compas_rhino.geometry.RhinoNurbsSurface.closest_point` to fix bug of rhino_curve to rhino_surface, plus return tuple instead.
 * Changed `compas_plotters.plotter.Plotter` to normal class instead of singleton.
 * Moved functionality of `compas.utilities.coercion` to `compas.data`.
+* Fixed bug in `compas.geometry.NurbsSurface.to_triangles()`.
 
 ### Removed